requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.0
marshmallow==3.21.1
openai==1.30.1
requests==2.32.2
